Moved render_field templatetag testcase to utils
